Clean up code. No functionality change expected.
